+++ gosh/params.py
@@ -9,6 +9,98 @@ import re
 #
 ########################
 
+# **************************************************************
+# 1. experment type related definitions
+def mode():
+    return 1
+    # parameter to change assimilation mode
+    # runoff ensembles will change accordingly.
+    # 1: Earth2Obs, 2: ERA20CM, 3: VIC_BC, 4: -25% baised (ELSE_KIM2009/E2O/ERA20CM)
+
+def conflag():
+    return 3
+    # converstion flag for observation converstions 
+    #  1 - Directly values 
+    #  2 - Anomalies
+    #  3 - Normalized values
+    #  4 - Log converted values
+
+def mapname():
+    return "amz_06min"
+    # realted CaMa-Flood map directory
+    # [e.g. : glb_15min, glb_06min, Mkg_06min, etc.]
+    # Check 
+
+def map_dimension():
+    fname=CaMa_dir()+"/map/"+mapname()+"/params.txt"
+    f=open(fname,"r")
+    lines=f.readlines()
+    f.close()
+    #-------
+    nx     = int(filter(None, re.split(" ",lines[0]))[0])
+    ny     = int(filter(None, re.split(" ",lines[1]))[0])
+    gsize  = float(filter(None, re.split(" ",lines[3]))[0])
+    return nx,ny,gsize
+
+# **************************************************************
+# 2. Data assimilation related definitions
+def patch_size():
+    #return 0
+    return 100
+    # the size of the local patch of LETKF(Local ** EnKF)
+    # 0: only 1 pixel (the pixel itself) belongs to its local patch
+    # 100: empirical local patch
+
+def DA_dir():
+    return "/cluster/data6/menaka/HydroDA"
+    # directory of HydroDA
+    # where src, dat, sat, out exsits
+
+def patch_dir():
+    return "/cluster/data6/menaka/Empirical_LocalPatch/local_patch"
+    #return "/cluster/data6/menaka/covariance/local_patch"
+    #return "/cluster/data6/menaka/covariance/local_patchMS"
+    #return "/cluster/data6/menaka/covariance/local_patch_0.80"
+
+def patch_name():
+    return "amz_06min_S14FD"
+
+def patch_id():
+    return "0.60"
+
+def thersold():
+    return 0.60
+    # thersold to define the local patch
+
+def initial_infl():
+    return 1.08
+    # initial inflation parameter
+
+def rho():
+    return -1.0
+    # -1.0 : adaptive inflation will be used as in Myoshi et al (2011)
+    # positive : fixed inflation parameter will be used
+
+def sigma_b():
+    return 0.0400000
+    # bacground variance of inflation for adaptive inflation Myoshi et al (2011)
+
+def ens_mem(mode=mode()):
+    if mode == 1:
+        return 21
+    
+    if mode == 2:
+        return 20
+
+    if mode == 3:
+        return 20
+
+    if mode == 3:
+        return 20
+    # number of ensemble members
+
+# **************************************************************
+# 3. Experiment timimngs 
 def timestep():
     return 86400 # outer timestep in seconds
 
@@ -18,6 +110,17 @@ def starttime():
 def endtime():
     return (2004,1,1) # end date: [year,month,date]
                       # *note: this date is not included
+
+# **************************************************************
+# 4. Spinup options
+def spinup_mode():
+    return 0
+    # 0: do spinup simulation for both (corrupted and true) simulation
+    # 1: do spin up only at corrupted simulation
+    # 2: do spin up only at true simulation
+    # 3: no spinup simulation at all
+    ### if initial restart file is ready, spinup simulation is no need
+
 def spinup_end_year():
     return 2001
 
@@ -27,12 +130,8 @@ def spinup_end_month():
 def spinup_end_date():
     return 31
 
-def mode():
-    return 3
-    # parameter to change assimilation mode
-    # runoff ensembles will change accordingly.
-    # 1: Earth2Obs, 2: ERA20CM, 3: VIC_BC, 4: -25% baised (ELSE_KIM2009/E2O/ERA20CM)
-
+# **************************************************************
+# 5. Runoff forcing 
 def runname(num):
     if num == 1:
         return "E2O"
@@ -66,20 +165,6 @@ def experiment():
     f.close()
     return exp
     
-def ens_mem(mode=mode()):
-    if mode == 1:
-        return 21
-    
-    if mode == 2:
-        return 20
-
-    if mode == 3:
-        return 20
-
-    if mode == 3:
-        return 20
-    # number of ensemble members
-
 def max_lat():
     return 80. # maximum latitude of assimilation
                # *note: SWOT ovservation is not available beyond 80 degs. this should be less or equal to 80
@@ -139,13 +224,6 @@ def assimE():
     # data Assimilation's Region (East Edge at latitude)
     # *note: should be smaller or equal to 170
 
-def patch_size():
-    #return 0
-    return 100
-    # the size of the local patch of LETKF(Local ** EnKF)
-    # 0: only 1 pixel (the pixel itself) belongs to its local patch
-    # 100: empirical local patch
-
 def err_expansion():
     return 1.0
     # variance-covariance expansion
@@ -187,68 +265,8 @@ def CaMa_dir():
     # directory of CaMa-Flood
     # indicate the directory of ./map or ./src and other folders
 
-def mapname():
-    return "amz_06min"
 
-def map_dimension():
-    fname=CaMa_dir()+"/map/"+mapname()+"/params.txt"
-    f=open(fname,"r")
-    lines=f.readlines()
-    f.close()
-    #-------
-    nx     = int(filter(None, re.split(" ",lines[0]))[0])
-    ny     = int(filter(None, re.split(" ",lines[1]))[0])
-    gsize  = float(filter(None, re.split(" ",lines[3]))[0])
-    return nx,ny,gsize
 
-def DA_dir():
-    return "/cluster/data6/menaka/HydroDA"
-    # directory of HydroDA
-    # where src, dat, sat, out exsits
-
-def patch_dir():
-    return "/cluster/data6/menaka/Empirical_LocalPatch/local_patch"
-    #return "/cluster/data6/menaka/covariance/local_patch"
-    #return "/cluster/data6/menaka/covariance/local_patchMS"
-    #return "/cluster/data6/menaka/covariance/local_patch_0.80"
-
-def patch_name():
-    return "amz_06min_S14FD"
-
-def patch_id():
-    return "0.60"
-
-def spinup_mode():
-    return 0
-    # 0: do spinup simulation for both (corrupted and true) simulation
-    # 1: do spin up only at corrupted simulation
-    # 2: do spin up only at true simulation
-    # 3: no spinup simulation at all
-    ### if initial restart file is ready, spinup simulation is no need
-
-def ovs_err():
-    return 0.100000
-    # size of SWOT observation error in meters
-    # 10cm for 1km^2 water body or 25cm < 1km^2 water area
-    # should be at least 0.02
-    # hope to be below 0.10
-
-def thersold():
-    return 0.60
-    # thersold to define the local patch
-
-def initial_infl():
-    return 1.08
-    # initial inflation parameter
-
-def rho():
-    return -1.0
-    # -1.0 : adaptive inflation will be used as in Myoshi et al (2011)
-    # positive : fixed inflation parameter will be used
-
-def sigma_b():
-    return 0.0400000
-    # bacground variance of inflation for adaptive inflation Myoshi et al (2011)
 
 def MKLdir():
     return "/opt/intel/compilers_and_libraries_2016.3.170/mac/mkl"
@@ -263,7 +281,13 @@ def output_er():
     # 0 for saving & 1 for deleting
     # those files may be more than 400GB, so erasing is recommended if not necessary
 
+def obs_name():
+    return "HydroWeb"
+
 def HydroWeb_dir():
+    return "/cluster/data6/menaka/HydroWeb"
+
+def obs_dir():
     return "/cluster/data6/menaka/HydroWeb"
 
 def make_log():
