+++ gosh/run_mool.sh
@@ -23,8 +23,8 @@
 ################################################################################################
 
 ### SET "mool PBS" @ IIS U-Tokyo
-#PBS -q F40
-#PBS -l select=1:ncpus=40:mem=60gb
+#PBS -q E20
+#PBS -l select=1:ncpus=20:mem=60gb
 #PBS -l place=scatter
 #PBS -j oe
 #PBS -m ea
@@ -43,7 +43,7 @@ which python
 
 # get number of cpus
 #export NCPUS=`cat ${PBS_NODEFILE} | wc -l`
-NCPUS=40
+NCPUS=20
 
 # OMP Settings
 export OMP_NUM_THREADS=$NCPUS
@@ -59,8 +59,9 @@ cd $HydroDA
 # experiment : edit the experiment name in here it will be written in $HydroDA/$EXP/exp.txt
 # before running run_mool.sh , please edit the necessary experimental settings in params.py
 
-EXP="VIC_BC_HydroWeb11"
+# EXP="VIC_BC_HydroWeb11"
 #EXP="E2O_HydroWeb23"
+EXP="test_wse"
 #IFACTOR="1.08"
 
 mkdir -p $HydroDA"/out/"$EXP
@@ -72,12 +73,14 @@ echo $EXP > $HydroDA"/out/"$EXP"/exp.txt"
 echo $NCPUS > $HydroDA"/out/"$EXP"/ncpus.txt"
 
 # copy params.py
-cp -r "$HydroDA/gosh/params.py" "$HydroDA/out/$EXP/params.py"
+cp -r $HydroDA/gosh/params.py      $HydroDA/out/$EXP/params.py
 #cp -r "$HydroDA/gosh/params.py" "$HydroDA/src/params.py"
 
-cp -r $HydroDA"/src/"run.py $HydroDA"/out/"$EXP"/"run.py
-cp -r $HydroDA"/src/"main_code.py $HydroDA"/out/"$EXP"/"main_code.py
-cp -r $HydroDA"/src/"prep_init.py $HydroDA"/out/"$EXP"/"prep_init.py
+cp -r $HydroDA/src/run.py         $HydroDA/out/$EXP/run.py
+cp -r $HydroDA/src/main_code.py   $HydroDA/out/$EXP/main_code.py
+cp -r $HydroDA/src/prep_init.py   $HydroDA/out/$EXP/prep_init.py
+cp -r $HydroDA/src/prep_runoff.py $HydroDA/out/$EXP/prep_runoff.py
+cp -r $HydroDA/src/prep_obs.py    $HydroDA/out/$EXP/prep_obs.py
 
 cd $HydroDA"/out/"$EXP
